backend/requirements.txt
# comfy
torch
torchsde
torchvision
torchaudio
einops
transformers
tokenizers>=0.13.3
sentencepiece
safetensors
psutil
spandrel
spandrel_extra_arches

# fastapi
fastapi
uvicorn
python-socketio
pydantic

tinydb
numpy
pillow
sympy
tqdm
packaging

mpmath
setuptools
scipy
filelock
filetype
urllib3
fsspec
requests
pyyaml
starlette
pydantic
httpx
aiotinydb
aiofiles
loguru
psutil
